Refactor solvers into interface-based system
