+++ Assets/Scripts/Pathfinding/IPathSolver.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1368a048ad574ccd861feb4c1717b83a
